Move api back to the old location
